% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoKin.R
\name{calculate_service_level}
\alias{calculate_service_level}
\title{Calculate service level}
\usage{
calculate_service_level(demand_data, product)
}
\arguments{
\item{demand_data}{A data frame containing demand data with columns: Product, Demand.}

\item{product}{The product for which service level is to be calculated.}
}
\value{
The calculated service level.
}
\description{
This function calculates the service level for a given product based on demand data.
}
