% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoKin.R
\name{stockin}
\alias{stockin}
\title{Stockin function}
\usage{
stockin(
  parameters_df,
  inventory_data,
  demand_data,
  orderleadtime,
  delivery_leadtime,
  product_service_levels,
  order_leadtime_filter = NULL
)
}
\arguments{
\item{parameters_df}{A data frame containing product and store information.}

\item{inventory_data}{A data frame containing inventory levels for each product and store.}

\item{demand_data}{A data frame containing demand data with columns: Product, Month, Demand.}

\item{orderleadtime}{Lead time for placing orders.}

\item{delivery_leadtime}{Lead time for delivery of orders.}

\item{product_service_levels}{A data frame containing service levels for each product.}

\item{order_leadtime_filter}{Filter for specific order lead time.}
}
\value{
A data frame with order information including product, store, reorder status, order quantity, and colisage.
}
\description{
This function calculates reorder quantities based on demand, inventory, and service level parameters.
}
